refactor: util 파일 분리
