+++ src/__tests__/productManager.test.ts
@@ -60,7 +60,7 @@ describe('상품 관리 도메인 테스트', () => {
     };
 
     vendingMachineProductManager.addProduct(newProduct);
-    vendingMachineProductManager.editProduct(targetProductName, targetProduct);
+    vendingMachineProductManager.editProduct(targetProduct);
 
     const { name, price, quantity } =
       vendingMachineProductManager.getProducts()[0];
