+++ src/ts/components/NavigatorComponent.ts
@@ -1,5 +1,15 @@
-import { $, $$, on } from '../dom';
+import { $, $$ } from '../dom';
 import { MAIN_PAGE } from '../constants';
+import { on } from '../utils';
+
+const render = (clickedTab: string, tab: string) => {
+  $(`.nav__${tab}-button`).className =
+    tab === clickedTab
+      ? `nav__button nav__${tab}-button nav__button--focused`
+      : `nav__button nav__${tab}-button`;
+  $(`.${tab}-section`).className =
+    tab === clickedTab ? `${tab}-section` : `${tab}-section hide`;
+};
 
 export default class NavigatorComponent {
   private $navList = $('.nav__list');
@@ -22,7 +32,7 @@ export default class NavigatorComponent {
 
   private changeTab = (targetTab) => {
     Array.from(this.$tabs).forEach((tabElement: HTMLElement) => {
-      this.render(targetTab, tabElement.dataset.tab);
+      render(targetTab, tabElement.dataset.tab);
     });
   };
 
@@ -30,13 +40,4 @@ export default class NavigatorComponent {
     const targetTab = window.location.pathname;
     this.changeTab(targetTab.replace('/', '') || MAIN_PAGE);
   };
-
-  private render(clickedTab: string, tab: string) {
-    $(`.nav__${tab}-button`).className =
-      tab === clickedTab
-        ? `nav__button nav__${tab}-button nav__button--focused`
-        : `nav__button nav__${tab}-button`;
-    $(`.${tab}-section`).className =
-      tab === clickedTab ? `${tab}-section` : `${tab}-section hide`;
-  }
 }
