+++ src/ts/dom.ts
@@ -1,42 +1,4 @@
-import { SNACK_BAR_DELAY_TIME, ERROR_MESSAGE } from './constants';
-
-export const on = (
-  target: Window | HTMLElement,
-  eventName: string,
-  callback
-) => {
-  target.addEventListener(eventName, callback);
-};
-
-export const emit = (
-  target: Window | HTMLElement,
-  eventName: string,
-  detail = {}
-) => {
-  const customEvent = new CustomEvent(eventName, detail);
-  target.dispatchEvent(customEvent);
-};
-
-const generateSnackBarTemplate = (message: string): string => `
-  <div class="snack-bar-container__snack-bar--error">
-    <p class="snack-bar-container__message">${message}</p>
-  </div>
-`;
-
-export const renderSnackBar = (
-  container: HTMLElement,
-  message: string
-): void => {
-  container.insertAdjacentHTML('beforeend', generateSnackBarTemplate(message));
-
-  setTimeout(() => {
-    const $snackBar = document.querySelector(
-      '.snack-bar-container__snack-bar--error'
-    );
-    $snackBar.classList.add('hide');
-    container.removeChild($snackBar);
-  }, SNACK_BAR_DELAY_TIME);
-};
+import { ERROR_MESSAGE } from './constants';
 
 export const $ = (selector: string): HTMLElement =>
   document.querySelector(selector);
